Removed the ElGamalManage completely. Test has been passed
