+++ ElGamalExt/ElGamal.cs
@@ -18,31 +18,154 @@ using System.Security.Cryptography;
 using System.Text;
 using System.Xml.Linq;
 using System.Numerics;
+using BigIntegerExt;
 
 namespace ElGamalExt
 {
-    /*public enum ElGamalPaddingMode : byte
+    public class ElGamal : AsymmetricAlgorithm
     {
-        ANSIX923,
-        LeadingZeros,
-        TrailingZeros,
-        BigIntegerPadding
-    }*/
+		private ElGamalKeyStruct o_key_struct;
+		public ElGamalKeyStruct KeyStruct
+		{
+			get
+			{
+				if (NeedToGenerateKey())
+				{
+					CreateKeyPair(KeySizeValue);
+				}
+				return o_key_struct;
+			}
+			set
+			{
+				o_key_struct = value;
+			}
+		}
 
-    public abstract class ElGamal : AsymmetricAlgorithm
-    {
-        //public ElGamalPaddingMode Padding;
+		public ElGamal()
+		{
+			// create the key struct and set all of the big integers to zero
+			o_key_struct = new ElGamalKeyStruct
+			{
+				P = BigInteger.Zero,
+				G = BigInteger.Zero,
+				Y = BigInteger.Zero,
+				X = BigInteger.Zero
+			};
 
-        public abstract void ImportParameters(ElGamalParameters p_parameters);
-        public abstract ElGamalParameters ExportParameters(bool p_include_private_params);
-        public abstract byte[] EncryptBigInteger(BigInteger p_data);
-        public abstract BigInteger DecryptBigInteger(byte[] p_data);
-        public abstract byte[] Sign(byte[] p_hashcode);
-        public abstract bool VerifySignature(byte[] p_hashcode, byte[] p_signature);
+			// set the default key size value
+			KeySizeValue = 384;
 
-        public abstract byte[] Multiply(byte[] p_first, byte[] p_second);
+			// set the default padding mode
+			//Padding = ElGamalPaddingMode.BigIntegerPadding;
 
-        public override string ToXmlString(bool p_include_private)
+			// set the range of legal keys
+			LegalKeySizesValue = new[] { new KeySizes(384, 1088, 8) };
+		}
+
+		private bool NeedToGenerateKey()
+		{
+			return (o_key_struct.P == 0) && (o_key_struct.G == 0) && (o_key_struct.Y == 0);
+		}
+
+		private void CreateKeyPair(int p_key_strength)
+		{
+			using (var x_random_generator = RandomNumberGenerator.Create())
+			{
+				// create the large prime number, P
+				o_key_struct.P = o_key_struct.P.GenPseudoPrime(p_key_strength, 16, x_random_generator);
+
+				// create the two random numbers, which are smaller than P
+				o_key_struct.X = new BigInteger();
+				o_key_struct.X = o_key_struct.X.GenRandomBits(p_key_strength - 1, x_random_generator);
+				o_key_struct.G = new BigInteger();
+				o_key_struct.G = o_key_struct.G.GenRandomBits(p_key_strength - 1, x_random_generator);
+
+				// compute Y
+				o_key_struct.Y = BigInteger.ModPow(o_key_struct.G, o_key_struct.X, o_key_struct.P);
+
+				//o_key_struct.Padding = Padding;
+			}
+		}
+
+		public byte[] EncryptData(BigInteger p_data)
+		{
+			if (NeedToGenerateKey())
+			{
+				CreateKeyPair(KeySizeValue);
+			}
+
+			using (var x_enc = new ElGamalEncryptor(o_key_struct))
+			{
+				return x_enc.ProcessBigInteger(p_data);
+			}
+		}
+
+		public BigInteger DecryptData(byte[] p_data)
+		{
+			if (NeedToGenerateKey())
+			{
+				CreateKeyPair(KeySizeValue);
+			}
+
+			var x_enc = new ElGamalDecryptor(o_key_struct);
+
+			return x_enc.ProcessByteBlock(p_data);
+		}
+
+		public byte[] Multiply(byte[] p_first, byte[] p_second)
+		{
+			return Homomorphism.ElGamalHomomorphism.Multiply(p_first, p_second, o_key_struct.P.ToByteArray());
+		}
+
+		public void ImportParameters(ElGamalParameters p_parameters)
+		{
+			// obtain the  big integer values from the byte parameter values
+			o_key_struct.P = new BigInteger(p_parameters.P);
+			o_key_struct.G = new BigInteger(p_parameters.G);
+			o_key_struct.Y = new BigInteger(p_parameters.Y);
+			//o_key_struct.Padding = p_parameters.Padding;
+
+			if (p_parameters.X != null && p_parameters.X.Length > 0)
+			{
+				o_key_struct.X = new BigInteger(p_parameters.X);
+			}
+
+			// set the length of the key based on the import
+			KeySizeValue = o_key_struct.P.BitCount();
+			//Padding = o_key_struct.Padding;
+		}
+
+		public ElGamalParameters ExportParameters(bool p_include_private_params)
+		{
+			if (NeedToGenerateKey())
+			{
+				CreateKeyPair(KeySizeValue);
+			}
+
+			// create the parameter set and set the public values of the parameters
+			var x_params = new ElGamalParameters
+			{
+				P = o_key_struct.P.ToByteArray(),
+				G = o_key_struct.G.ToByteArray(),
+				Y = o_key_struct.Y.ToByteArray(),
+				//Padding = o_key_struct.Padding
+			};
+
+			// if required, include the private value, X
+			if (p_include_private_params)
+			{
+				x_params.X = o_key_struct.X.ToByteArray();
+			}
+			else
+			{
+				// ensure that we zero the value
+				x_params.X = new byte[1];
+			}
+
+			return x_params;
+		}
+
+		public override string ToXmlString(bool p_include_private)
         {
             var x_params = ExportParameters(p_include_private);
